Edit: moved authentication dtos and entities to domain layer
