+++ src/modules/authentication/domain/entities/IUserAuthentication.ts
