+++ src/modules/authentication/infra/controllers/useSignupController.ts
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useState } from 'react';
 
-import IUserSignupDTO from 'modules/authentication/dtos/IUserSignupDTO';
+import IUserSignupDTO from 'modules/authentication/domain/dtos/IUserSignupDTO';
 import makeSignUserService from 'modules/authentication/services/factories/makeSignUserService';
 import { useToastContext } from 'shared/view/contexts';
