+++ src/modules/authentication/infra/repositories/UsersRepository.ts
@@ -1,6 +1,6 @@
-import IUserLoginDTO from 'modules/authentication/dtos/IUserLoginDTO';
-import IUserSignupDTO from 'modules/authentication/dtos/IUserSignupDTO';
-import IUserAuthentication from 'modules/authentication/entities/IUserAuthentication';
+import IUserLoginDTO from 'modules/authentication/domain/dtos/IUserLoginDTO';
+import IUserSignupDTO from 'modules/authentication/domain/dtos/IUserSignupDTO';
+import IUserAuthentication from 'modules/authentication/domain/entities/IUserAuthentication';
 import IUsersRepository from 'modules/authentication/repositories/IUsersRepository';
 import IUser from 'shared/domain/entities/IUser';
 import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
