+++ src/modules/authentication/repositories/IUsersRepository.ts
@@ -1,6 +1,6 @@
-import IUserLoginDTO from 'modules/authentication/dtos/IUserLoginDTO';
-import IUserSignupDTO from 'modules/authentication/dtos/IUserSignupDTO';
-import IUserAuthentication from 'modules/authentication/entities/IUserAuthentication';
+import IUserLoginDTO from 'modules/authentication/domain/dtos/IUserLoginDTO';
+import IUserSignupDTO from 'modules/authentication/domain/dtos/IUserSignupDTO';
+import IUserAuthentication from 'modules/authentication/domain/entities/IUserAuthentication';
 import IUser from 'shared/domain/entities/IUser';
 
 export default interface IUsersRepository {
